backend/requirements.txt
# Core - always required
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
slowapi>=0.1.9
aiosmtplib>=3.0.2
jinja2>=3.1.4

# Conditional - ENABLE_TOTP_2FA
pyotp>=2.9.0

# Conditional - ENABLE_OAUTH2
httpx>=0.28.1

# Conditional - ENABLE_AUDIT_LOGGING
structlog>=24.4.0

# Dev / Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

boto3>=1.34
google-auth>=2.0
